+++ examples/example_1.jl
@@ -1,4 +1,4 @@
-using MembraneRD, Random, Colors
+using MembraneRD, Random, Colors, MembraneRD.Filters
 
 species = @species A B EA EB
 
@@ -99,7 +99,7 @@ saver = Pusher(Tuple{Float64,State})
 colors = [color("yellow"),color("blue"),color("black"),color("black")]/30
 stats = TimeFilter(ProgressShower(T), 
 #   Measurer(M; name="test_example_1", Nsave),
-#   StopWatchFilter(display ∘ Plotter(posx, posy; colors); seconds=1.0),
+    StopWatchFilter(display ∘ Plotter(posx, posy; colors); seconds=1.0),
     saver; times)
 @time run_RD!(s, M, T; stats, rng)
